backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.104.1
google-api-core==2.28.1
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.5.0
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
oauthlib==3.3.1
passlib==1.7.4
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0
pyparsing==3.2.5
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Pillow==10.1.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
redis==5.0.1
celery==5.3.4
